prevent trigger of systemd rate limiting
By default, the systemd will stop restarting the service if it fails too fast.

We had a situation where the node boots, but the dns resolution is not up yet, which
causes the service to fail in quick succession.

Prevent triggering of rate limiting by implementing a 10 second interval between retries.