make sure retry works in case of failure
The previous commit has implemented a 10 second sleep when autossh fails
to start due to dns resolution failure or any other problem.

I just thought that systemd may not interpret that as a failure, since
`sleep 10` will have exit code 0. So when that happens, we ensure to
send a non-zero exit code.